+++ tests/database/migrations/2021_11_15_000000_create_test_table.php
@@ -1,0 +1,38 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\Schema;
+use Php\Support\Laravel\Database\Schema\Postgres\Blueprint;
+
+class CreateTestTable extends Migration
+{
+    private static string $table = 'tests';
+
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create(
+            static::$table,
+            static function (Blueprint $table) {
+                $table->primaryUUID();
+                $table->string('name');
+                $table->boolean('enabled');
+                $table->timestamps();
+            }
+        );
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists(static::$table);
+    }
+}